% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{agemap_data_list}
\alias{agemap_data_list}
\title{AGEMAP Data List}
\format{
A list of data frames with rows for the samples and columns for the genes (Unigene ID).
}
\usage{
agemap_data_list
}
\description{
A list of datasets containing the pre-processed gene expression for aging in mice
from Muscle (Ms), Adrenal glands (Ad), Hippocampus (Hi), Spinal cord (Sp), and Spleen (S).
}
\references{
Zahn, Jacob M., Suresh Poosala, Art B. Owen, Donald K. Ingram, Ana Lustig, Arnell Carter, Ashani T. Weeraratna, et al. 2007. “AGEMAP: A Gene Expression Database for Aging in Mice.” PLoS Genetics 3 (11): e201. https://doi.org/10.1371/journal.pgen.0030201.
}
\keyword{datasets}
