% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{agemap_label_list}
\alias{agemap_label_list}
\title{AGEMAP Label List}
\format{
A list of vectors (1, 6, 16, 24 month) for the age labels of the samples.
}
\usage{
agemap_label_list
}
\description{
A list of vectors for the labels of \link{agemap_data_list}.
}
\references{
Zahn, Jacob M., Suresh Poosala, Art B. Owen, Donald K. Ingram, Ana Lustig, Arnell Carter, Ashani T. Weeraratna, et al. 2007. “AGEMAP: A Gene Expression Database for Aging in Mice.” PLoS Genetics 3 (11): e201. https://doi.org/10.1371/journal.pgen.0030201.
}
\keyword{datasets}
