% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_function.R
\name{mc.permute.test}
\alias{mc.permute.test}
\title{Permutation test for min-MCC, MCTC, and post-hoc MCMI analysis}
\usage{
mc.permute.test(
  study.data.list,
  study.label.list,
  w.est = NULL,
  n.perm = 1000,
  n.parallel = 50
)
}
\arguments{
\item{study.data.list}{a list of data frames or matrices with rows for the genes and columns for the samples.}

\item{study.label.list}{a list of vectors of the class labels.}

\item{w.est}{a vector of pre-defined class weights that add up to 1. If not specified, it will be calculated from the data.}

\item{n.perm}{permutation times}

\item{n.parallel}{number of cores used for parallel computation}
}
\value{

}
\description{
Permutation test for min-MCC, MCTC, and post-hoc MCMI analysis
}
