% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_function.R
\name{mc.stat}
\alias{mc.stat}
\title{A function to combine the MCTC and minMCC}
\usage{
mc.stat(study.data.list, study.label.list, w.est = NULL)
}
\arguments{
\item{study.data.list}{a list of data frames or matrices with rows for the genes and columns for the samples.}

\item{study.label.list}{a list of vectors of the class labels.}

\item{w.est}{a vector of pre-defined class weights that add up to 1. If not specified, it will be calculated from the data.}
}
\value{

}
\description{
A function to combine the MCTC and minMCC
}
\keyword{internal}
