% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_function.R
\name{MCTC}
\alias{MCTC}
\title{Multi-class Total Correlation (MC-TC)}
\usage{
MCTC(study.data.list, study.label.list, w.est = NULL)
}
\arguments{
\item{study.data.list}{a list of data frames or matrices with rows for the genes and columns for the samples.}

\item{study.label.list}{a list of vectors of the class labels.}

\item{w.est}{a vector of pre-defined class weights that add up to 1. If not specified, it will be calculated from the data.}
}
\value{
Multi-class total correlation (MC-TC) for multiple studies.
}
\description{
This function is used to calculate the multi-class total correlation (MC-TC) among multiple studies of gene expression, where the gene expressions are assumed to follow Gaussian distribution.
}
\references{
Lu, Shuya, Jia Li, Chi Song, Kui Shen, and George C. Tseng. 2010. “Biomarker Detection in the Integration of Multiple Multi-Class Genomic Studies.” Bioinformatics  26 (3): 333–40. https://doi.org/10.1093/bioinformatics/btp669.
}
